table-styles renamed to table-styled-components removed all string litterals and replaced with actual styled component
