+++ components/styled/icon/src/organisms/icon-button.tsx
@@ -15,14 +15,14 @@ const StyledText = styled.span`
   margin-left: 0.4rem;
 `
 
-const IconButton: FC<Props> = ({ iconProps, text, ...rest }) => (
+const IconButton: FC<IconButtonProps> = ({ iconProps, text, ...rest }: IconButtonProps) => (
   <StyledInvisibleButton {...rest} role="button">
     <Icon {...iconProps} />
     {text && <StyledText>{text}</StyledText>}
   </StyledInvisibleButton>
 )
 
-interface Props extends ButtonHTMLAttributes<HTMLButtonElement> {
+export interface IconButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   iconProps: IconProps
   text?: string
 }
